+++ src/gpsea/analysis/temporal/__init__.py
