+++ src/gpsea/analysis/temporal/_api.py
