+++ src/gpsea/analysis/temporal/_base.py
