+++ src/gpsea/analysis/temporal/endpoint/__init__.py
