+++ src/gpsea/analysis/temporal/endpoint/_impl.py
