+++ src/gpsea/analysis/temporal/stats/__init__.py
