+++ src/gpsea/analysis/temporal/stats/_api.py
