+++ src/gpsea/analysis/temporal/stats/_impl.py
